[AAE-11695] move version compatibility directive
