+++ lib/content-services/src/lib/content-metadata/services/content-type-property.service.spec.ts
@@ -17,12 +17,13 @@
 
 import { TestBed } from '@angular/core/testing';
 import { ContentTypePropertiesService } from './content-type-property.service';
-import { CardViewItem, CardViewSelectItemModel, CardViewTextItemModel, setupTestBed, VersionCompatibilityService } from '@alfresco/adf-core';
+import { CardViewItem, CardViewSelectItemModel, CardViewTextItemModel, setupTestBed } from '@alfresco/adf-core';
 import { ContentTestingModule } from '../../testing/content.testing.module';
 import { TranslateModule } from '@ngx-translate/core';
 import { ContentTypeService } from '../../content-type';
 import { of } from 'rxjs';
 import { Node, TypeEntry } from '@alfresco/js-api';
+import { VersionCompatibilityService } from '../../version-compatibility/version-compatibility.service';
 
 describe('ContentTypePropertyService', () => {
