+++ lib/content-services/src/lib/version-compatibility/public-api.ts
@@ -15,11 +15,6 @@
  * limitations under the License.
  */
 
-export * from './highlight.directive';
-export * from './logout.directive';
-export * from './node-download.directive';
-export * from './upload.directive';
-export * from './tooltip-card/tooltip-card.directive';
-export * from './infinite-select-scroll.directive';
-
-export * from './directive.module';
+export * from './version-compatibility.directive';
+export * from './version-compatibility.service';
+export * from './version-compatibility.module';
