+++ lib/content-services/src/lib/version-compatibility/version-compatibility.service.ts
@@ -19,7 +19,7 @@ import { Injectable } from '@angular/core';
 import { filter } from 'rxjs/operators';
 import { ReplaySubject } from 'rxjs';
 import { VersionInfo, RepositoryInfo } from '@alfresco/js-api';
-import { DiscoveryApiService } from './discovery-api.service';
+import { DiscoveryApiService } from '@alfresco/adf-core';
 
 @Injectable({
     providedIn: 'root'
