+++ lib/content-services/src/public-api.ts
@@ -37,5 +37,6 @@ export * from './lib/aspect-list/index';
 export * from './lib/content-type/index';
 export * from './lib/new-version-uploader';
 export * from './lib/interfaces/index';
+export * from './lib/version-compatibility/index';
 
 export * from './lib/content.module';
