+++ lib/core/src/lib/services/public-api.ts
@@ -61,7 +61,6 @@ export * from './ecm-user.service';
 export * from './identity-user.service';
 export * from './identity-group.service';
 export * from './identity-role.service';
-export * from './version-compatibility.service';
 export * from './auth-bearer.interceptor';
 export * from './oauth2.service';
 export * from './language.service';
